Rename custom object
