+++ docs/CustomObject.md
@@ -1,6 +1,6 @@
 # Custom Object
 
-[The custom object example](../force-app/main/default/objects/Visit__c) demonstrates the capability to work with custom objects in the Mobile Offline app. Like standard objects, users can view, create, and edit records of custom objects while their device is offline.
+[The custom object example](../force-app/main/default/objects/StarterKitCustomObject__c) demonstrates the capability to work with custom objects in the Mobile Offline app. Like standard objects, users can view, create, and edit records of custom objects while their device is offline.
 
 The following steps generally describe the setup of a custom object for use in offline scenarios. You can see a specific example of these steps in the "Configure" section below.
 - Create a custom object and the associated fields needed for your use case.
@@ -14,37 +14,37 @@ The following steps generally describe the setup of a custom object for use in o
 
 ## Configure Custom Object
 
-In this example, a `Visit` custom object is configured by deploying the relevant code artifacts to your org.
+In this example, a `StarterKitCustomObject` custom object is configured by deploying the relevant code artifacts to your org.
 
-1. Deploy the following code artifacts for the `Visit` object, in the order outlined below.
+1. Deploy the following code artifacts for the `StarterKitCustomObject` object, in the order outlined below.
 
     ```sh
-    # Create the Visit object with 5 new fields
-    sfdx force:source:deploy -p force-app/main/default/objects/Visit__c
+    # Create the StarterKitCustomObject object with 5 new fields
+    sfdx force:source:deploy -p force-app/main/default/objects/StarterKitCustomObject__c
 
     # Add LWCs
-    sfdx force:source:deploy -p force-app/main/default/lwc/viewVisitRecord
-    sfdx force:source:deploy -p force-app/main/default/lwc/createVisitRecord
-    sfdx force:source:deploy -p force-app/main/default/lwc/editVisitRecord
+    sfdx force:source:deploy -p force-app/main/default/lwc/viewStarterKitCustomObjectRecord
+    sfdx force:source:deploy -p force-app/main/default/lwc/createStarterKitCustomObjectRecord
+    sfdx force:source:deploy -p force-app/main/default/lwc/editStarterKitCustomObjectRecord
 
     # Add quick actions
-    sfdx force:source:deploy -p force-app/main/default/quickActions/Visit__c.view.quickAction-meta.xml
-    sfdx force:source:deploy -p force-app/main/default/quickActions/Visit__c.create.quickAction-meta.xml
-    sfdx force:source:deploy -p force-app/main/default/quickActions/Visit__c.edit.quickAction-meta.xml
+    sfdx force:source:deploy -p force-app/main/default/quickActions/StarterKitCustomObject__c.view.quickAction-meta.xml
+    sfdx force:source:deploy -p force-app/main/default/quickActions/StarterKitCustomObject__c.create.quickAction-meta.xml
+    sfdx force:source:deploy -p force-app/main/default/quickActions/StarterKitCustomObject__c.edit.quickAction-meta.xml
 
-    # Add quick actions to the Visit Layout
-    sfdx force:source:deploy -p "force-app/main/default/layouts/Visit__c-Visit Layout.layout-meta.xml"
+    # Add quick actions to the StarterKitCustomObject Layout
+    sfdx force:source:deploy -p "force-app/main/default/layouts/StarterKitCustomObject__c-StarterKitCustomObject Layout.layout-meta.xml"
 
-    # Add the Visits tab for user to create records on the browser
-    sfdx force:source:deploy -p force-app/main/default/tabs/Visit__c.tab-meta.xml
+    # Add the StarterKitCustomObjects tab for user to create records on the browser
+    sfdx force:source:deploy -p force-app/main/default/tabs/StarterKitCustomObject__c.tab-meta.xml
 
-    # Add the permission set "Offline Custom Objects" for accessing Visit object
+    # Add the permission set "Offline Custom Objects" for accessing StarterKitCustomObject object
     sfdx force:source:deploy -p force-app/main/default/permissionsets/Offline_Custom_Objects.permissionset-meta.xml
     ``` 
 
 2. In your org, an org admin should assign the permission set `Offline Custom Objects` to org users.
 
-3. [Configure your Offline Briefcase](../README.md#define-an-offline-briefcase) to include the `Visit` object. 
-    - Make sure to create a few `Visit` records to be included in the briefcase.
+3. [Configure your Offline Briefcase](../README.md#define-an-offline-briefcase) to include the `StarterKitCustomObject` object. 
+    - Make sure to create a few `StarterKitCustomObject` records to be included in the briefcase.
 
-4. Log in to the Mobile Offline App. You should see the new `Visit` object!
+4. Log in to the Mobile Offline App. You should see the new `StarterKitCustomObject` object!
