+++ force-app/main/default/layouts/StarterKitCustomObject__c-StarterKitCustomObject Layout.layout-meta.xml
@@ -72,17 +72,17 @@
     <platformActionList>
         <actionListContext>Record</actionListContext>
         <platformActionListItems>
-            <actionName>Visit__c.view</actionName>
+            <actionName>StarterKitCustomObject__c.view</actionName>
             <actionType>QuickAction</actionType>
             <sortOrder>0</sortOrder>
         </platformActionListItems>
         <platformActionListItems>
-            <actionName>Visit__c.edit</actionName>
+            <actionName>StarterKitCustomObject__c.edit</actionName>
             <actionType>QuickAction</actionType>
             <sortOrder>1</sortOrder>
         </platformActionListItems>
         <platformActionListItems>
-            <actionName>Visit__c.create</actionName>
+            <actionName>StarterKitCustomObject__c.create</actionName>
             <actionType>QuickAction</actionType>
             <sortOrder>2</sortOrder>
         </platformActionListItems>
