+++ force-app/main/default/lwc/createStarterKitCustomObjectRecord/createStarterKitCustomObjectRecord.css
