+++ force-app/main/default/lwc/createStarterKitCustomObjectRecord/createStarterKitCustomObjectRecord.js-meta.xml
@@ -2,9 +2,10 @@
 <LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
     <apiVersion>56.0</apiVersion>
     <isExposed>true</isExposed>
-    <masterLabel>View Visit Record</masterLabel>
-    <description>Component to view a single Visit. Will also display any draft items in accordion if any drafts exist for the record.</description>
+    <masterLabel>Create StarterKitCustomObject Record</masterLabel>
+    <description>Creates a new StarterKitCustomObject.</description>
     <targets>
+        <target>lightning__GlobalAction</target>
         <target>lightning__RecordPage</target>
         <target>lightning__RecordAction</target>
     </targets>
