+++ force-app/main/default/lwc/editStarterKitCustomObjectRecord/__tests__/data/getRecord.json
