+++ force-app/main/default/lwc/editStarterKitCustomObjectRecord/editStarterKitCustomObjectRecord.css
