+++ force-app/main/default/lwc/editStarterKitCustomObjectRecord/editStarterKitCustomObjectRecord.js
@@ -1,14 +1,14 @@
 import { LightningElement, api, wire } from "lwc";
 import { getRecord } from "lightning/uiRecordApi";
 
-import NAME_FIELD from "@salesforce/schema/Visit__c.Name";
-import START_TIME__C_FIELD from "@salesforce/schema/Visit__c.Start_Time__c";
-import END_TIME__C_FIELD from "@salesforce/schema/Visit__c.End_Time__c";
-import PRIORITY__C_FIELD from "@salesforce/schema/Visit__c.Priority__c";
-import STATUS__C_FIELD from "@salesforce/schema/Visit__c.Status__c";
-import ADDRESS__C_FIELD from "@salesforce/schema/Visit__c.Address__c";
+import NAME_FIELD from "@salesforce/schema/StarterKitCustomObject__c.Name";
+import START_TIME__C_FIELD from "@salesforce/schema/StarterKitCustomObject__c.Start_Time__c";
+import END_TIME__C_FIELD from "@salesforce/schema/StarterKitCustomObject__c.End_Time__c";
+import PRIORITY__C_FIELD from "@salesforce/schema/StarterKitCustomObject__c.Priority__c";
+import STATUS__C_FIELD from "@salesforce/schema/StarterKitCustomObject__c.Status__c";
+import ADDRESS__C_FIELD from "@salesforce/schema/StarterKitCustomObject__c.Address__c";
 
-export default class EditVisit__cRecord extends LightningElement {
+export default class EditStarterKitCustomObject__cRecord extends LightningElement {
   @api recordId;
   @api objectApiName;
 
@@ -32,7 +32,7 @@ export default class EditVisit__cRecord extends LightningElement {
   }
 
   onSuccess(event) {
-    console.log("Updated Visit__c", event.detail);
+    console.log("Updated StarterKitCustomObject__c", event.detail);
     // Dismiss modal on success
     this.dismiss(event);
   }
