+++ force-app/main/default/lwc/editStarterKitCustomObjectRecord/editStarterKitCustomObjectRecord.js-meta.xml
