+++ force-app/main/default/lwc/viewStarterKitCustomObjectRecord/__tests__/data/getRecord.json
@@ -1,7 +1,7 @@
 {
     "fields": {
         "Name": {
-            "value": "Mock Visit"
+            "value": "Mock StarterKitCustomObject"
         }
     }
  }
