+++ force-app/main/default/lwc/viewStarterKitCustomObjectRecord/viewStarterKitCustomObjectRecord.css
