+++ force-app/main/default/lwc/viewStarterKitCustomObjectRecord/viewStarterKitCustomObjectRecord.html
@@ -5,7 +5,7 @@
                 
                 <div class="title-slot">
                     <c-record-header record-name={name} object-api-name={objectApiName}
-                        object-label="Visit" icon-name="custom:custom9">
+                        object-label="StarterKitCustomObject" icon-name="custom:custom9">
                     </c-record-header>
                 </div>
