+++ force-app/main/default/objects/StarterKitCustomObject__c/fields/End_Time__c.field-meta.xml
