+++ force-app/main/default/objects/StarterKitCustomObject__c/fields/Status__c.field-meta.xml
