+++ force-app/main/default/quickActions/StarterKitCustomObject__c.create.quickAction-meta.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <QuickAction xmlns="http://soap.sforce.com/2006/04/metadata">
     <actionSubtype>ScreenAction</actionSubtype>
-    <label>Edit</label>
-    <lightningWebComponent>editVisitRecord</lightningWebComponent>
+    <label>Create</label>
+    <lightningWebComponent>createStarterKitCustomObjectRecord</lightningWebComponent>
     <optionsCreateFeedItem>false</optionsCreateFeedItem>
     <type>LightningWebComponent</type>
-    <icon>editActionIcon</icon>
 </QuickAction>
