+++ force-app/main/default/quickActions/StarterKitCustomObject__c.edit.quickAction-meta.xml
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <QuickAction xmlns="http://soap.sforce.com/2006/04/metadata">
     <actionSubtype>ScreenAction</actionSubtype>
-    <label>View</label>
-    <lightningWebComponent>viewVisitRecord</lightningWebComponent>
+    <label>Edit</label>
+    <lightningWebComponent>editStarterKitCustomObjectRecord</lightningWebComponent>
     <optionsCreateFeedItem>false</optionsCreateFeedItem>
     <type>LightningWebComponent</type>
+    <icon>editActionIcon</icon>
 </QuickAction>
