+++ force-app/main/default/quickActions/StarterKitCustomObject__c.view.quickAction-meta.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <QuickAction xmlns="http://soap.sforce.com/2006/04/metadata">
     <actionSubtype>ScreenAction</actionSubtype>
-    <label>Create</label>
-    <lightningWebComponent>createVisitRecord</lightningWebComponent>
+    <label>View</label>
+    <lightningWebComponent>viewStarterKitCustomObjectRecord</lightningWebComponent>
     <optionsCreateFeedItem>false</optionsCreateFeedItem>
     <type>LightningWebComponent</type>
 </QuickAction>
