+++ force-app/main/default/tabs/StarterKitCustomObject__c.tab-meta.xml
